requirements.txt
fastapi==0.65.2
uvicorn==0.13.4
pydantic==1.8.2
pydantic[email]
google-cloud-storage
python-dotenv